chatbot-pro/pyproject.toml
[project]
name = "ra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "langchain>=1.2.1",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.3",
    "pdfplumber>=0.11.9",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
]
